python-app/requirements.txt
alembic
python-chess
mysql-connector-python
pandas
tqdm
